requirements.txt
asgiref==3.11.1
attrs==25.4.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
command-not-found==0.3
crispy-bootstrap5==2025.6
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==5.2.11
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
filelock==3.6.0
greenlet==3.0.3
httplib2==0.20.2
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
pipenv==11.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.11
pycairo==1.20.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
PyYAML==5.4.1
referencing==0.37.0
rpds-py==0.30.0
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.28
sqlparse==0.5.5
systemd-python==234
termcolor==2.4.0
typing_extensions==4.10.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
zipp==1.0.0
